Move class for tests
